+++ src/server.ts
@@ -1,15 +1,9 @@
 import Fastify, { FastifyRequest, FastifyReply } from "fastify";
-import { MinesweeperGame } from "./game";
-import { renderBoardText } from "./serverBoardText";
-import { GameConfig, Position, GameStatus } from "./types";
-import { randomUUID } from "crypto";
 import { readFileSync } from "fs";
 import { join } from "path";
-import { routeSchemas } from "./schemas";
-import { CellState } from "./types";
-
 // Use require for the MCP plugin to work around module resolution issues
 const mcpPlugin = require("@mcp-it/fastify");
+import routesPlugin from "./routes";
 
 const fastify = Fastify({ logger: true });
 
@@ -23,25 +17,6 @@ const VERSION = packageJson.version;
 // - PORT: Server port (default: 8080)
 // - ADMIN_KEY: Admin key for protected endpoints
 const PORT = process.env.PORT || 8080;
-const ADMIN_KEY = process.env.ADMIN_KEY;
-
-// Authorization middleware
-// Fastify preHandler hook for admin authentication
-const requireAdminKey = async (req: FastifyRequest, reply: FastifyReply) => {
-  const adminKey = req.headers["x-admin-key"] || req.headers["authorization"];
-
-  if (!adminKey || adminKey !== ADMIN_KEY) {
-    reply.code(401).send({
-      error: "Unauthorized",
-      message:
-        "Valid admin key required. Provide it via 'X-Admin-Key' or 'Authorization' header.",
-    });
-    return;
-  }
-};
-
-// In-memory game sessions: { [id]: MinesweeperGame }
-const games: Record<string, MinesweeperGame> = {};
 
 // Serve index.html at root - inline HTML to avoid file path issues
 fastify.get("/", async (req: FastifyRequest, reply: FastifyReply) => {
@@ -183,288 +158,9 @@ curl -X POST http://localhost:${PORT}/api/game/[game-id]/reveal \\
   `);
 });
 
-// Health check endpoint with proper schema
-fastify.get(
-  "/api/health",
-  {
-    schema: routeSchemas.health,
-  },
-  async (req: FastifyRequest, reply: FastifyReply) => {
-    try {
-      reply.status(200).send({
-        status: "ok",
-        timestamp: new Date().toISOString(),
-        uptime: process.uptime(),
-        version: VERSION,
-      });
-    } catch (error) {
-      reply.status(500).send({
-        status: "error",
-        message: "Health check failed",
-      });
-    }
-  },
-);
-
-// Create new game endpoint with proper schema
-fastify.post<{ Body: GameConfig }>(
-  "/api/game",
-  {
-    schema: routeSchemas.createGame,
-  },
-  async (req: FastifyRequest<{ Body: GameConfig }>, reply: FastifyReply) => {
-    const config = req.body;
-
-    // Validate configuration
-    const errors: string[] = [];
-
-    if (!config.width || config.width < 3 || config.width > 50) {
-      errors.push("Width must be between 3 and 50");
-    }
-    if (!config.height || config.height < 3 || config.height > 50) {
-      errors.push("Height must be between 3 and 50");
-    }
-    if (
-      !config.bombPercentage ||
-      config.bombPercentage < 1 ||
-      config.bombPercentage > 99
-    ) {
-      errors.push("Bomb percentage must be between 1 and 99");
-    }
-
-    if (errors.length > 0) {
-      return reply
-        .code(400)
-        .send({ error: "Invalid configuration", details: errors });
-    }
-
-    const gameId = randomUUID();
-    const game = new MinesweeperGame(config);
-    games[gameId] = game;
-
-    const gameState = game.getGameState();
-    reply.code(201).send({
-      id: gameId,
-      config: game.getConfig(),
-      status: gameState.status,
-      board: renderBoardText(game.getConfig(), gameState),
-      flags: gameState.grid
-        .flat()
-        .filter((cell) => cell.state === CellState.FLAGGED).length,
-      remainingCells: gameState.totalSafeCells - gameState.revealedCells,
-    });
-  },
-);
-
-// Get game state endpoint with proper schema
-fastify.get<{ Params: { id: string } }>(
-  "/api/game/:id",
-  {
-    schema: routeSchemas.getGameState,
-  },
-  async (
-    req: FastifyRequest<{ Params: { id: string } }>,
-    reply: FastifyReply,
-  ) => {
-    const game = games[req.params.id];
-    if (!game) return reply.code(404).send({ error: "Game not found" });
-
-    const gameState = game.getGameState();
-    reply.send({
-      id: req.params.id,
-      config: game.getConfig(),
-      status: gameState.status,
-      board: renderBoardText(game.getConfig(), gameState),
-      flags: gameState.grid
-        .flat()
-        .filter((cell) => cell.state === CellState.FLAGGED).length,
-      remainingCells: gameState.totalSafeCells - gameState.revealedCells,
-    });
-  },
-);
-
-// Reveal cell endpoint with proper schema
-fastify.post<{ Params: { id: string }; Body: Position }>(
-  "/api/game/:id/reveal",
-  {
-    schema: routeSchemas.revealCell,
-  },
-  async (
-    req: FastifyRequest<{ Params: { id: string }; Body: Position }>,
-    reply: FastifyReply,
-  ) => {
-    const game = games[req.params.id];
-    if (!game) return reply.code(404).send({ error: "Game not found" });
-
-    const { row, col } = req.body as Position;
-
-    // Validate position
-    const config = game.getConfig();
-    if (row < 0 || row >= config.height || col < 0 || col >= config.width) {
-      return reply.code(400).send({ error: "Invalid position" });
-    }
-
-    const success = game.revealCell({ row, col });
-    if (!success) {
-      return reply.code(400).send({
-        error: "Cannot reveal cell - it may already be revealed or flagged",
-      });
-    }
-
-    const gameState = game.getGameState();
-    reply.send({
-      id: req.params.id,
-      config: game.getConfig(),
-      status: gameState.status,
-      board: renderBoardText(game.getConfig(), gameState),
-      flags: gameState.grid
-        .flat()
-        .filter((cell) => cell.state === CellState.FLAGGED).length,
-      remainingCells: gameState.totalSafeCells - gameState.revealedCells,
-    });
-  },
-);
-
-// Flag cell endpoint with proper schema
-fastify.post<{ Params: { id: string }; Body: Position }>(
-  "/api/game/:id/flag",
-  {
-    schema: routeSchemas.flagCell,
-  },
-  async (
-    req: FastifyRequest<{ Params: { id: string }; Body: Position }>,
-    reply: FastifyReply,
-  ) => {
-    const game = games[req.params.id];
-    if (!game) return reply.code(404).send({ error: "Game not found" });
-
-    const { row, col } = req.body as Position;
-
-    // Validate position
-    const config = game.getConfig();
-    if (row < 0 || row >= config.height || col < 0 || col >= config.width) {
-      return reply.code(400).send({ error: "Invalid position" });
-    }
-
-    const success = game.toggleFlag({ row, col });
-    if (!success) {
-      return reply
-        .code(400)
-        .send({ error: "Cannot flag cell - it may already be revealed" });
-    }
-
-    const gameState = game.getGameState();
-    reply.send({
-      id: req.params.id,
-      config: game.getConfig(),
-      status: gameState.status,
-      board: renderBoardText(game.getConfig(), gameState),
-      flags: gameState.grid
-        .flat()
-        .filter((cell) => cell.state === CellState.FLAGGED).length,
-      remainingCells: gameState.totalSafeCells - gameState.revealedCells,
-    });
-  },
-);
-
-// List games endpoint (admin only) with proper schema
-fastify.get(
-  "/api/games",
-  {
-    schema: routeSchemas.listGames,
-    preHandler: requireAdminKey,
-  },
-  async (req: FastifyRequest, reply: FastifyReply) => {
-    reply.send({
-      ids: Object.keys(games),
-      total: Object.keys(games).length,
-      message: "Active game sessions",
-    });
-  },
-);
-
-// Command endpoint with proper schema
-fastify.post<{ Params: { id: string }; Body: { command: string } }>(
-  "/api/game/:id/command",
-  {
-    schema: routeSchemas.executeCommand,
-  },
-  async (
-    req: FastifyRequest<{ Params: { id: string }; Body: { command: string } }>,
-    reply: FastifyReply,
-  ) => {
-    const game = games[req.params.id];
-    if (!game) return reply.code(404).send({ error: "Game not found" });
-
-    const input = req.body.command.trim();
-    let message = "";
-
-    try {
-      const parts = input.toLowerCase().split(/\s+/);
-      const action = parts[0];
-      const row = parseInt(parts[1], 10);
-      const col = parseInt(parts[2], 10);
-
-      if (isNaN(row) || isNaN(col)) {
-        return reply.code(400).send({
-          error: "Invalid command format. Use: 'reveal 3 4' or 'flag 2 5'",
-        });
-      }
-
-      const config = game.getConfig();
-      if (row < 0 || row >= config.height || col < 0 || col >= config.width) {
-        return reply.code(400).send({ error: "Position out of bounds" });
-      }
-
-      let success = false;
-      switch (action) {
-        case "reveal":
-        case "r":
-          success = game.revealCell({ row, col });
-          message = success
-            ? `Revealed cell at (${row}, ${col})`
-            : `Cannot reveal cell at (${row}, ${col}) - it may already be revealed or flagged`;
-          break;
-        case "flag":
-        case "f":
-          success = game.toggleFlag({ row, col });
-          message = success
-            ? `Toggled flag at (${row}, ${col})`
-            : `Cannot flag cell at (${row}, ${col}) - it may already be revealed`;
-          break;
-        default:
-          return reply
-            .code(400)
-            .send({ error: "Unknown command. Use 'reveal' or 'flag'" });
-      }
-
-      if (!success) {
-        return reply.code(400).send({ error: message });
-      }
-
-      const gameState = game.getGameState();
-      reply.send({
-        message,
-        gameState: {
-          id: req.params.id,
-          config: game.getConfig(),
-          status: gameState.status,
-          board: renderBoardText(game.getConfig(), gameState),
-          flags: gameState.grid
-            .flat()
-            .filter((cell) => cell.state === CellState.FLAGGED).length,
-          remainingCells: gameState.totalSafeCells - gameState.revealedCells,
-        },
-      });
-    } catch (error) {
-      reply.code(400).send({ error: (error as Error).message });
-    }
-  },
-);
-
 async function startServer() {
   try {
-    // Register the MCP plugin with proper configuration
+    // Register the MCP plugin FIRST
     await fastify.register(mcpPlugin, {
       name: "Minesweeper Game Server",
       description:
@@ -473,6 +169,9 @@ async function startServer() {
       addDebugEndpoint: true,
     });
 
+    // Register routes AFTER MCP plugin
+    await fastify.register(routesPlugin);
+
     const address = await fastify.listen({
       port: Number(PORT),
       host: "0.0.0.0",
@@ -480,11 +179,11 @@ async function startServer() {
 
     fastify.log.info(`Server listening at ${address}`);
     fastify.log.info(`MCP SSE server available at ${address}/mcp/sse`);
-    fastify.log.info(`MCP debug endpoint available at ${address}/mcp/debug`);
+    fastify.log.info(`MCP debug endpoint available at ${address}/mcp/tools`);
     console.log(`\n🎮 Minesweeper Server Ready!`);
     console.log(`📍 Web UI: ${address}`);
     console.log(`🔌 MCP Endpoint: ${address}/mcp/sse`);
-    console.log(`🐛 Debug Tools: ${address}/mcp/debug`);
+    console.log(`🐛 Debug Tools: ${address}/mcp/tools`);
   } catch (err) {
     fastify.log.error(err);
     process.exit(1);
